requirements.txt
pygame~=2.5.2
numpy~=1.26.4
pyaudio~=0.2.11
torch~=2.4.0
openai-whisper~=20231117
pandas~=2.2.2
matplotlib~=3.9.2
pyyaml~=6.0.2
seaborn~=0.13.2
faster-whisper~=1.0.3
vosk~=0.3.45
jiwer~=3.0.3
tokenizers~=0.19.1
webrtcvad~=2.0.10
plotly~=5.24.1
pydub~=0.25.1
scipy~=1.14.0
tabulate~=0.9.0